loop body
so far I've specified requirements of the matching method, chosen to use a map for men and a map for women (each map will contain the person's ID # as an int to identify them with (the key), as well as their associaited preference list, represnted in an array, ordered from highest preferred to least preferred (the value)). The "free" men will be iterated through until there are none left, with a clause declaring the the set of them decreases over time to ensure termination of the loop. Then we search in the preference list of the current man, find the closest preferred woman, and check whether or not she has already been matched. If not, she is matched to the current man.